Add parameter to compareFront function in parser
Add tokenStart parameter to compareFront function, so that two arrays of block-level tokens can be compared more precisely. If the tokenStart parameter is defined, the function will only compare the two array up to the index of tokenStart. If tokenStart isn't defined, the function will compare the two arrays for the full length of the frontTokens parameter.

Replace redundant code in the codeBlock rule with the compareFront function.